Fixed soe recurity issues

+++ src/vscp/common/vscpremotetcpif.cpp
@@ -1951,7 +1951,7 @@ VscpRemoteTcpIf::readLevel2Registers(uint32_t reg,
     if (lastpageCnt) nPages++;
     unsigned long allRcvValue = (unsigned long)pow(2.0, nPages) - 1;
 
-    unsigned long resendTime = m_registerOpResendTimeout * (1 + nPages);
+    unsigned long resendTime = (unsigned long)m_registerOpResendTimeout * (1 + nPages);
     uint32_t startTime = vscp_getMsTimeStamp();
 
     /*
